Extract PluralUtils.js from IString.js
Move plural rule engine (plurals_default, loadPlurals, _fncs) out of
IString.js into a standalone PluralUtils.js module. IString retains
backward-compatible aliases for all three static members so no callers
need to change.

Also add docs/IString-Refactoring.md with the full split plan.
